Initial commit for tensorflow/python/keras to Github project keras-team/keras.
Currently the code is still under development/migration and it is NOT the source of truth. The code in tensorflow/tensorflow repository IS the source of truth for the moment. Please send any PR to that. We will send announcement for the formal switch when we finish all the migration/workflow work.

We will add more documents and setup the developer workflow in the near future.

PiperOrigin-RevId: 345532569