Clarify documentation of DepthwiseConv2D
As discussed in [tensorflow issue 49933](https://github.com/tensorflow/tensorflow/issues/49933).

The documentation for `DepthwiseConv2D` describes the operation as: split input into individual channels, convolve each with the layer's kernel and finally stack the results. This gives the impression the same kernel is applied to all channels. But in the implementation tensorflow implementation as well as the [MobileNet Paper](https://arxiv.org/pdf/1704.04861.pdf) (see page 3 formula 3), a different kernel is applied per channel.