Add 3 experimental backend API in preparing for RNG behavior change.
The 3 new APIs are used to enable/disable/check the usage of `tf.random.Generator`, which will be the new backend for all the RNG in Keras. We plan to switch on the new code path by default in tf 2.8, and the behavior change will like to cause some breakage on user side (eg if the test is checking against some golden nubmer). These 3 APIs will allow user to disable and switch back to legacy behavior if they prefer. In future (eg tf 2.10), we expect to totally remove the legacy code path (stateful random Ops), and these 3 APIs will be removed as well.

PiperOrigin-RevId: 414769654