Create a class that delegates LossScaleOptimizer creation to the right subclass depending on the inner optimizer's type.
To avoid name confusion and massive code changes, we are creating a BaseLossScaleOptimizer class that delegates object creation to LSO or LSOV3 based on the type of inner optimizer. The BaseLossScaleOptimizer will be exported as tf.keras.mixed_precision.LossScaleOptimizer so that type check on tf.keras.mixed_precision.LossScaleOptimizer would pass for both LossScaleOptimizer and LossScaleOptimizerV3.

PiperOrigin-RevId: 423184717