Remove tf1 symbol export for augmentation image layers
Image preprocessing layers are intended as a TF2 API, and image augmentation
layers do not support training/inference call time branching in a TF1
friendly way.

We do not believe we have usage of these layers in TF1, and would like to
discourage future users attempting to user augmentation layers with the
incompatible TF1 base model class, so we are removing the symbol export.

PiperOrigin-RevId: 428924959