Move the _distributed_container attribute from ResourceVariable to handle.
ResourceVariable is now a CompositeTensor, but can't be packed and unpacked like other CompositeTensor. One blocker is when a ResourceVariable is reconstructed from a dt_resource handle, it will lose the _distributed_container attribute. Because this attribute is added to ResourceVariable after its construction and not all ResourceVariables have this attribute. Moving the attribute from ResourceVariable to handle so this attribute will persistent through packing and unpacking cycle.

PiperOrigin-RevId: 479367785