Broadcast mask to input during normalization.
This commit changes the way that masked grouped normalization is computed. Previously, the mask had to have the same shape as the input. This is not always the case, for example when the mask is a single channel image and the input is a three channel image. This change allows the mask to have any shape, and the input will be broadcast to match the shape of the mask. This makes it possible to use masked grouped normalization in a wider variety of cases.

PiperOrigin-RevId: 538097599