Support torch tensor inputs to data adapter (#399)
We use the `ARRAY_TYPES` in the epoch iterator to decide if an input
tensor to decide if an input type should be handle by the
`ArrayDataAdapter`. Without this change was seeing some errors passing
torch tensors to fit/predict on the torch backend.

Because this means that torch will be imported on all backends, we need
to move out import hack to all backends as well.