Reduce redundancy on saving model (#18871)
* Fix saving_api.py for merging changelog of model.save()

I took this code out of the .keras model saving logic. It is more reasonable for checking file overwriting each of saving types that .keras, .h5, .hdf5 . And When comparing the two versions, this code is more recent version.

* Fix method docstring and warning description of keras.saving.save_model() and model.save()

In deprectaion warning, It notice only the case of calling model.save(). so I added the case of keras.saving.save_model
And method docstring too. Additionally, I added deprecation warning in Args section for save_format.

* Remove redundancy of model.save()

* Update test assersion in test_h5_deprecation_warning() of saving_api_test.py